converts TestBookPage.js and css into BookingPage.js and css
